app/requirements.txt
openai
streamlit-confetti
